+++ src/validators/validation-types/index.js
@@ -1,2 +1,7 @@
 export { requiredField } from './requiredField.validator'
 export { textField } from './textField.validator'
+export { emailField } from './emailField.validator'
+export { checkboxField } from './checkboxField.validator'
+export { radioField } from './radioField.validator'
+export { selectField } from './selectField.validator'
+export { textRequired } from './textRequired.validator'